restructured images folder to have additional folder for sample products
