+++ client/src/assets/images/sampleProducts/art2.png
