+++ client/src/assets/images/sampleProducts/bag.jpg
