+++ client/src/assets/images/sampleProducts/gaming-keyboard.png
