+++ client/src/assets/images/sampleProducts/house-lamp.png
