+++ client/src/assets/images/sampleProducts/lace-wall.png
