+++ client/src/assets/images/sampleProducts/office-chair.png
