+++ client/src/assets/images/sampleProducts/red-dress.png
