+++ client/src/data/productData.json
@@ -3,7 +3,7 @@
     "id": 1,
     "name": "Test Product 1",
     "description": "1 Apparel Description Lorem ipsum dolor sit amet consectetur adipisicing elit. Lorem ipsum dolor sit amet consectetur adipisicing elit. Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/red-dress.png",
+    "img": "../src/assets/images/sampleProducts/red-dress.png",
     "price": 24.99,
     "rating": 4,
     "category": "Apparel"
@@ -12,7 +12,7 @@
     "id": 2,
     "name": "Test Product 2",
     "description": "2 Apparel Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/bag.jpg",
+    "img": "../src/assets/images/sampleProducts/bag.jpg",
     "price": 39.95,
     "rating": 5,
     "category": "Apparel"
@@ -21,7 +21,7 @@
     "id": 3,
     "name": "Test Product 3",
     "description": "3 Sports & Outdoors Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/brazuca-ball.png",
+    "img": "../src/assets/images/sampleProducts/brazuca-ball.png",
     "price": 24.99,
     "rating": 3,
     "category": "Sports & Outdoors"
@@ -30,7 +30,7 @@
     "id": 4,
     "name": "Test Product 4",
     "description": "4 Sports & Outdoors Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/baseball-glove.jpeg",
+    "img": "../src/assets/images/sampleProducts/baseball-glove.jpeg",
     "price": 39.95,
     "rating": 5,
     "category": "Sports & Outdoors"
@@ -39,7 +39,7 @@
     "id": 5,
     "name": "Test Product 5",
     "description": "5 Kids Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/kids-toy1.png",
+    "img": "../src/assets/images/sampleProducts/kids-toy1.png",
     "price": 24.99,
     "rating": 1,
     "category": "Kids"
@@ -48,7 +48,7 @@
     "id": 6,
     "name": "Test Product 6",
     "description": "6 Baby Care Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/baby-cleaning.png",
+    "img": "../src/assets/images/sampleProducts/baby-cleaning.png",
     "price": 39.95,
     "rating": null,
     "category": "Baby Care"
@@ -57,7 +57,7 @@
     "id": 7,
     "name": "Test Product 7",
     "description": "7 Baby Care Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/baby-care2.png",
+    "img": "../src/assets/images/sampleProducts/baby-care2.png",
     "price": 24.99,
     "rating": 5,
     "category": "Baby Care"
@@ -66,7 +66,7 @@
     "id": 8,
     "name": "Test Product 8",
     "description": "8 Home Decor Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/house-lamp.png",
+    "img": "../src/assets/images/sampleProducts/house-lamp.png",
     "price": 39.95,
     "rating": 3,
     "category": "Home Decor"
@@ -75,7 +75,7 @@
     "id": 9,
     "name": "Test Product 9",
     "description": "9 Home Decor Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/lace-wall.png",
+    "img": "../src/assets/images/sampleProducts/lace-wall.png",
     "price": 10.98,
     "rating": 1,
     "category": "Home Decor"
@@ -84,7 +84,7 @@
     "id": 10,
     "name": "Test Product 10",
     "description": "10 Art Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/art1.jpg",
+    "img": "../src/assets/images/sampleProducts/art1.jpg",
     "price": 10.98,
     "rating": 1,
     "category": "Art"
@@ -93,7 +93,7 @@
     "id": 11,
     "name": "Test Product 10",
     "description": "11 Art Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/art2.png",
+    "img": "../src/assets/images/sampleProducts/art2.png",
     "price": 10.98,
     "rating": 1,
     "category": "Art"
@@ -102,7 +102,7 @@
     "id": 12,
     "name": "Test Product 10",
     "description": "12 Gaming Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/gaming-keyboard.png",
+    "img": "../src/assets/images/sampleProducts/gaming-keyboard.png",
     "price": 10.98,
     "rating": 1,
     "category": "Gaming"
@@ -111,7 +111,7 @@
     "id": 13,
     "name": "Test Product 13",
     "description": "13 Office Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/office-chair.png",
+    "img": "../src/assets/images/sampleProducts/office-chair.png",
     "price": 10.98,
     "rating": 1,
     "category": "Office"
@@ -120,7 +120,7 @@
     "id": 14,
     "name": "Test Product 14",
     "description": "14 Apparel Test Description Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    "img": "../src/assets/images/shirt.jpg",
+    "img": "../src/assets/images/sampleProducts/shirt.jpg",
     "price": 39.95,
     "rating": 5,
     "category": "Apparel"
