+++ client/src/pages/Profile/Buyer/Profile/Accordion/ReviewsImgList.jsx
@@ -13,7 +13,7 @@ import { useEffect } from 'react';
 import { useLazyQuery } from '@apollo/client';
 import { User } from '../../../../../utils/queries';
 import Auth from '../../../../../utils/auth';
-import TestItem1 from '../../../../../assets/images/bag.jpg';
+import TestItem1 from '../../../../../assets/images/sampleProducts/bag.jpg';
 import { Rating } from '@mui/material';
 
 const style = {
